Changes from xpilot-cpp.

+++ src/client/clientmap.cpp
@@ -87,7 +87,7 @@ double Fuel_by_pos(int x, int y)
 
 int Target_by_index(int ind, int *xp, int *yp, int *dead_time, double *damage)
 {
-    if (ind < 0 || ind >= clMap.targets.size())
+    if (ind < 0 || ind >= (int)clMap.targets.size())
         return -1;
     target_t &target = clMap.targets[ind];
     *xp = target.pos / Setup->y;
@@ -123,7 +123,7 @@ int Target_alive(int x, int y, double *damage)
 
 int Handle_fuel(int ind, double fuel)
 {
-    if (ind < 0 || ind >= clMap.fuels.size())
+    if (ind < 0 || ind >= (int)clMap.fuels.size())
     {
         warn("Bad fuelstation index (%d)", ind);
         return -1;
@@ -165,7 +165,7 @@ int Cannon_dead_time_by_pos(int x, int y, int *dot)
 
 int Handle_cannon(int ind, int dead_time)
 {
-    if (ind < 0 || ind >= clMap.cannons.size())
+    if (ind < 0 || ind >= (int)clMap.cannons.size())
     {
         warn("Bad cannon index (%d)", ind);
         return 0;
@@ -176,7 +176,7 @@ int Handle_cannon(int ind, int dead_time)
 
 int Handle_target(int ind, int dead_time, double damage)
 {
-    if (ind < 0 || ind >= clMap.targets.size())
+    if (ind < 0 || ind >= (int)clMap.targets.size())
     {
         warn("Bad target index (%d)", ind);
         return 0;
@@ -239,7 +239,7 @@ int Handle_base(int id, int ind)
 {
     int i;
 
-    if (ind < 0 || ind >= clMap.bases.size())
+    if (ind < 0 || ind >= (int)clMap.bases.size())
     {
         warn("Bad homebase index (%d)", ind);
         return -1;
